Rewrite of dual-tree
This is essentially a complete rewrite of the package formerly known
as dubl-tree (now known as dual-tree) to be much simpler and more
regular.  It is now more directly based on the idea of a free algebra;
for motivation see my paper "Monoids: Theme and Variations"
(http://www.cis.upenn.edu/~byorgey/publications.html).

Trees can also now store incidental data in internal nodes, which we
can use for things like grouping transparent objects or other sorts of
"annotations" which should be placed at an internal node. To go along
with this, the fold function is much simpler and should now make sense
to expose to backends to give them more control over rendering. (See
https://github.com/diagrams/diagrams-core/issues/19).