add access to internal d nodes in foldDUAL
I am not sure why exactly I left this out before.  I think I convinced
myself it was the "right" thing to do.  To be sure, one does need to
be careful not to mix up the internal d values and the combined d
values provided at the leaves.  But that no longer seems to me a good
reason to deny the internal d values to the user.  And besides, we
need the internal d values, in order to convert DUALTrees into
something more structured which we can hand to diagrams backends!